Move repository implementations for model classes to dedicated package.
