--- mico-core/src/main/java/io/github/ust/mico/core/DependsOnRepository.java
@@ -1,6 +1,0 @@
-package io.github.ust.mico.core;
-
-import org.springframework.data.neo4j.repository.Neo4jRepository;
-
-public interface DependsOnRepository extends Neo4jRepository<DependsOn, Long> {
-}
