--- mico-core/src/main/java/io/github/ust/mico/core/ServiceInterfaceRepository.java
@@ -1,6 +1,0 @@
-package io.github.ust.mico.core;
-
-import org.springframework.data.neo4j.repository.Neo4jRepository;
-
-public interface ServiceInterfaceRepository extends Neo4jRepository<ServiceInterface, Long> {
-}
